+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricRows/Infos/FixedInfo/OrderInfo.cs
@@ -4,7 +4,7 @@
 using osu.Game.Graphics.Sprites;
 using osu.Game.Rulesets.Karaoke.Objects;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Infos.FixedInfo
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricRows.Infos.FixedInfo
 {
     public class OrderInfo : OsuSpriteText
     {
