+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricRows/Lyrics/Carets/DrawableLyricSplitterCaret.cs
@@ -9,7 +9,7 @@ using osu.Framework.Graphics.Sprites;
 using osu.Game.Graphics;
 using osuTK;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Carets
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricRows.Lyrics.Carets
 {
     public class DrawableLyricSplitterCaret : CompositeDrawable, IDrawableCaret
     {
