+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricRows/Lyrics/Carets/DrawableTimeTagRecordCaret.cs
@@ -10,7 +10,7 @@ using osu.Game.Rulesets.Karaoke.Graphics.Shapes;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osuTK;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Carets
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricRows.Lyrics.Carets
 {
     public class DrawableTimeTagRecordCaret : CompositeDrawable, IDrawableCaret, IHasTimeTag
     {
