+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricRows/Lyrics/Carets/IDrawableCaret.cs
@@ -3,7 +3,7 @@
 
 using osu.Framework.Graphics;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Carets
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricRows.Lyrics.Carets
 {
     public interface IDrawableCaret : IDrawable
     {
