+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricRows/Lyrics/Carets/IHasCaretPosition.cs
@@ -3,7 +3,7 @@
 
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Carets
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricRows.Lyrics.Carets
 {
     public interface IHasCaretPosition
     {
